Move unsafe to unsafe.go
