+++ helpers.go
@@ -3,29 +3,8 @@ package moira
 import (
 	"bytes"
 	"math"
-	"reflect"
 	"time"
-	"unsafe"
 )
-
-// UnsafeBytesToString converts source to string without copying
-func UnsafeBytesToString(b []byte) string {
-	hdr := *(*reflect.SliceHeader)(unsafe.Pointer(&b))
-	return *(*string)(unsafe.Pointer(&reflect.StringHeader{
-		Data: hdr.Data,
-		Len:  hdr.Len,
-	}))
-}
-
-// UnsafeStringToBytes converts string to source without copying
-func UnsafeStringToBytes(s string) []byte {
-	header := *(*reflect.StringHeader)(unsafe.Pointer(&s))
-	return *(*[]byte)(unsafe.Pointer(&reflect.SliceHeader{
-		Data: header.Data,
-		Len:  header.Len,
-		Cap:  header.Len,
-	}))
-}
 
 // BytesScanner allows to scan for subslices separated by separator
 type BytesScanner struct {
